requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.5
PyMySQL==1.1.2
python-decouple==3.8
Pillow==10.1.0